// src/utils/enums/campaignStatus.ts
export enum CampaignStatusEnum {
  Draft = 'Draft',
  PendingApproval = 'PendingApproval',
  Active = 'Active',
  Completed = 'Completed',
  Scheduled = 'Scheduled',
}

export enum CampaignStatusExtended {
  All = 'All',
  Draft = 'Draft',
  PendingApproval = 'PendingApproval',
  Active = 'Active',
  Completed = 'Completed',
  Scheduled = 'Scheduled',
}